requirements.txt
# Kafka client for Python
confluent-kafka
numpy
scikit-learn
requests
python-dotenv
psycopg2-binary
SQLAlchemy
streamlit
pandas
pytest
pytest-cov
requests-mock